+++ src/App.js
@@ -2,12 +2,14 @@ import React, { useState } from 'react'
 import { Route, Switch } from 'react-router-dom'
 import Header from './Header'
 import BylawsText from './BylawsText'
-import QuestionsForm from './QuestionsForm'
+import QuestionPage from './QuestionPage'
 import Start from './Start'
 import styled from 'styled-components/macro'
 
 function App() {
   const [answers, setAnswers] = useState([])
+
+  const addAnswer = (newAnswer) => setAnswers({ ...answers, ...newAnswer })
 
   return (
     <StyledCenterOnDesktop>
@@ -18,7 +20,7 @@ function App() {
             <Start />
           </Route>
           <Route path="/questions">
-            <QuestionsForm onAddAnswer={addAnswer} />
+            <QuestionPage onAddAnswer={addAnswer} />
           </Route>
           <Route path="/bylawstext">
             <BylawsText answers={answers} />
@@ -27,10 +29,6 @@ function App() {
       </StyledMainContainer>
     </StyledCenterOnDesktop>
   )
-
-  function addAnswer(newAnswer) {
-    setAnswers({ ...answers, ...newAnswer })
-  }
 }
 
 const StyledMainContainer = styled.div`
